refactor: split agent k8s backend modules
