+++ services/api-rs/crates/centaur-sandbox-agent-k8s/src/backend/mod.rs
@@ -1,0 +1,256 @@
+use std::pin::Pin;
+use std::time::Duration;
+
+use async_trait::async_trait;
+use centaur_sandbox_core::{
+    ObservedSandbox, SandboxBackend, SandboxError, SandboxHandle, SandboxId, SandboxIo,
+    SandboxResult, SandboxSpec, SandboxStatus,
+};
+use k8s_openapi::api::apps::v1::Deployment;
+use k8s_openapi::api::core::v1::{ConfigMap, Pod, Service};
+use k8s_openapi::api::networking::v1::NetworkPolicy;
+use kube::api::{AttachParams, DeleteParams, ListParams, Patch, PatchParams, PostParams};
+use kube::{Api, Client, Error};
+use serde_json::json;
+use tokio::io::{AsyncRead, AsyncWrite};
+use tokio::time::{Instant, sleep};
+
+use crate::config::AgentSandboxConfig;
+use crate::resources::*;
+use crate::{BACKEND_NAME, MANAGED_BY_LABEL, MANAGED_BY_VALUE, crd};
+
+mod iron_proxy;
+
+#[derive(Clone)]
+pub struct AgentSandboxBackend {
+    client: Client,
+    config: AgentSandboxConfig,
+}
+
+impl AgentSandboxBackend {
+    pub fn new(client: Client, config: AgentSandboxConfig) -> Self {
+        Self { client, config }
+    }
+
+    pub async fn try_default(namespace: impl Into<String>) -> SandboxResult<Self> {
+        let client = Client::try_default()
+            .await
+            .map_err(|err| SandboxError::Backend(format!("create kube client: {err}")))?;
+        Ok(Self::new(client, AgentSandboxConfig::new(namespace)))
+    }
+
+    fn sandboxes(&self) -> Api<crd::Sandbox> {
+        Api::namespaced(self.client.clone(), &self.config.namespace)
+    }
+
+    fn pods(&self) -> Api<Pod> {
+        Api::namespaced(self.client.clone(), &self.config.namespace)
+    }
+
+    fn config_maps(&self) -> Api<ConfigMap> {
+        Api::namespaced(self.client.clone(), &self.config.namespace)
+    }
+
+    fn services(&self) -> Api<Service> {
+        Api::namespaced(self.client.clone(), &self.config.namespace)
+    }
+
+    fn network_policies(&self) -> Api<NetworkPolicy> {
+        Api::namespaced(self.client.clone(), &self.config.namespace)
+    }
+
+    fn deployments(&self) -> Api<Deployment> {
+        Api::namespaced(self.client.clone(), &self.config.namespace)
+    }
+
+    async fn get_sandbox(&self, id: &SandboxId) -> SandboxResult<Option<crd::Sandbox>> {
+        match self.sandboxes().get(id.as_str()).await {
+            Ok(sandbox) => Ok(Some(sandbox)),
+            Err(err) if is_not_found(&err) => Ok(None),
+            Err(err) => Err(map_kube_error("get sandbox", err)),
+        }
+    }
+
+    async fn get_pod(&self, id: &SandboxId) -> SandboxResult<Option<Pod>> {
+        match self.pods().get(id.as_str()).await {
+            Ok(pod) => Ok(Some(pod)),
+            Err(err) if is_not_found(&err) => Ok(None),
+            Err(err) => Err(map_kube_error("get sandbox pod", err)),
+        }
+    }
+
+    async fn patch_replicas(&self, id: &SandboxId, replicas: i32) -> SandboxResult<()> {
+        let params = PatchParams::apply(&self.config.field_manager);
+        let patch = Patch::Merge(json!({ "spec": { "replicas": replicas } }));
+        self.sandboxes()
+            .patch(id.as_str(), &params, &patch)
+            .await
+            .map(|_| ())
+            .map_err(|err| map_kube_error("patch sandbox replicas", err))
+    }
+
+    async fn wait_until_running(&self, id: &SandboxId) -> SandboxResult<()> {
+        let deadline = Instant::now() + self.config.ready_timeout;
+        loop {
+            match self.status(id).await? {
+                SandboxStatus::Running => return Ok(()),
+                SandboxStatus::Gone | SandboxStatus::Stopped => {
+                    return Err(SandboxError::NotReady(format!(
+                        "sandbox {} reached terminal state before running",
+                        id.as_str()
+                    )));
+                }
+                status if Instant::now() >= deadline => {
+                    return Err(SandboxError::NotReady(format!(
+                        "sandbox {} did not become running before timeout; latest status: {status:?}",
+                        id.as_str()
+                    )));
+                }
+                _ => sleep(Duration::from_millis(500)).await,
+            }
+        }
+    }
+
+    async fn attach_io(&self, id: &SandboxId) -> SandboxResult<SandboxIo> {
+        if self.status(id).await? != SandboxStatus::Running {
+            return Err(SandboxError::NotReady(format!(
+                "agent sandbox {} is not running",
+                id.as_str()
+            )));
+        }
+        let params = AttachParams::default()
+            .container(self.config.container_name.clone())
+            .stdin(true)
+            .stdout(true)
+            .stderr(true)
+            .tty(false)
+            .max_stdout_buf_size(1024 * 1024)
+            .max_stderr_buf_size(1024 * 1024);
+        let mut attached = self
+            .pods()
+            .attach(id.as_str(), &params)
+            .await
+            .map_err(|err| map_kube_error("attach sandbox pod", err))?;
+        let stdin = attached
+            .stdin()
+            .map(|stream| Box::pin(stream) as Pin<Box<dyn AsyncWrite + Send>>);
+        let stdout = attached
+            .stdout()
+            .map(|stream| Box::pin(stream) as Pin<Box<dyn AsyncRead + Send>>);
+        let stderr = attached
+            .stderr()
+            .map(|stream| Box::pin(stream) as Pin<Box<dyn AsyncRead + Send>>);
+        let stdin = stdin.ok_or_else(|| SandboxError::Io("stdin was not attached".to_owned()))?;
+        let stdout =
+            stdout.ok_or_else(|| SandboxError::Io("stdout was not attached".to_owned()))?;
+        let stderr =
+            stderr.ok_or_else(|| SandboxError::Io("stderr was not attached".to_owned()))?;
+        // Keep kube's attach process alive as long as the returned streams are in use.
+        Ok(SandboxIo::with_guard(stdin, stdout, stderr, attached))
+    }
+}
+
+#[async_trait]
+impl SandboxBackend for AgentSandboxBackend {
+    fn name(&self) -> &'static str {
+        BACKEND_NAME
+    }
+
+    async fn create(&self, spec: SandboxSpec) -> SandboxResult<SandboxHandle> {
+        let id = SandboxId::new(next_sandbox_name());
+        let resolved_iron_proxy = self.resolve_iron_proxy(&id, &spec)?;
+        if let Err(err) = self
+            .create_iron_proxy_resources(&id, resolved_iron_proxy.as_ref())
+            .await
+        {
+            let _ = self.delete_iron_proxy_resources(&id).await;
+            return Err(err);
+        }
+        let sandbox = build_agent_sandbox(&id, &spec, &self.config, resolved_iron_proxy.as_ref())?;
+        let create_result = self
+            .sandboxes()
+            .create(&PostParams::default(), &sandbox)
+            .await
+            .map_err(|err| map_kube_error("create sandbox", err));
+        if let Err(err) = create_result {
+            let _ = self.delete_iron_proxy_resources(&id).await;
+            return Err(err);
+        }
+        if let Err(err) = self.wait_until_running(&id).await {
+            let _ = self.stop(&id).await;
+            return Err(err);
+        }
+        Ok(SandboxHandle::new(id, BACKEND_NAME))
+    }
+
+    async fn open_io(&self, id: &SandboxId) -> SandboxResult<SandboxIo> {
+        self.attach_io(id).await
+    }
+
+    async fn status(&self, id: &SandboxId) -> SandboxResult<SandboxStatus> {
+        let Some(sandbox) = self.get_sandbox(id).await? else {
+            return Ok(SandboxStatus::Gone);
+        };
+        let replicas = sandbox.spec.replicas.unwrap_or(1);
+        let pod = self.get_pod(id).await?;
+        Ok(sandbox_status_from_pod(replicas, pod.as_ref()))
+    }
+
+    async fn observe(&self, id: &SandboxId) -> SandboxResult<ObservedSandbox> {
+        let status = self.status(id).await?;
+        Ok(ObservedSandbox::new(id.clone(), BACKEND_NAME, status))
+    }
+
+    async fn list_observed(&self) -> SandboxResult<Vec<ObservedSandbox>> {
+        let params =
+            ListParams::default().labels(&format!("{MANAGED_BY_LABEL}={MANAGED_BY_VALUE}"));
+        let sandboxes = self
+            .sandboxes()
+            .list(&params)
+            .await
+            .map_err(|err| map_kube_error("list sandboxes", err))?;
+        let mut observed = Vec::with_capacity(sandboxes.items.len());
+        for sandbox in sandboxes.items {
+            let Some(name) = sandbox.metadata.name else {
+                continue;
+            };
+            let id = SandboxId::new(name);
+            observed.push(self.observe(&id).await?);
+        }
+        Ok(observed)
+    }
+
+    async fn stop(&self, id: &SandboxId) -> SandboxResult<()> {
+        match self
+            .sandboxes()
+            .delete(id.as_str(), &DeleteParams::default())
+            .await
+        {
+            Ok(_) => Ok(()),
+            Err(err) if is_not_found(&err) => Ok(()),
+            Err(err) => Err(map_kube_error("delete sandbox", err)),
+        }?;
+        self.delete_iron_proxy_resources(id).await
+    }
+
+    async fn pause(&self, id: &SandboxId) -> SandboxResult<()> {
+        self.patch_replicas(id, 0).await
+    }
+
+    async fn resume(&self, id: &SandboxId) -> SandboxResult<()> {
+        self.patch_replicas(id, 1).await?;
+        self.wait_until_running(id).await
+    }
+}
+
+fn is_not_found(err: &Error) -> bool {
+    matches!(err, Error::Api(api_error) if api_error.code == 404)
+}
+
+fn map_kube_error(operation: &str, err: Error) -> SandboxError {
+    if is_not_found(&err) {
+        SandboxError::NotFound(operation.to_owned())
+    } else {
+        SandboxError::Backend(format!("{operation}: {err}"))
+    }
+}
