+++ services/api-rs/crates/centaur-sandbox-agent-k8s/src/tests.rs
@@ -1,8 +1,12 @@
 use std::collections::BTreeMap;
 
 use centaur_iron_proxy::SourcePolicy;
-use centaur_sandbox_core::{CredentialProfile, MountKind, ResourceLimits, SandboxSpec};
+use centaur_sandbox_core::{
+    CredentialProfile, MountKind, ResourceLimits, SandboxId, SandboxSpec, SandboxStatus,
+};
 use k8s_openapi::api::core::v1::{Pod, PodCondition, PodStatus};
+
+use crate::resources::*;
 
 use super::*;
